Add files via upload
Self-Reflection
This visit opened my eyes to many things about the sewage system. To be honest, I used to not understand how sewage system works at all especially the part how sewage is treated and how it is safe to be released into river. But thanks to this visit, I genuinely got few insights and closer understanding the process behind the system. While not truly perfect, it is indeed a working system.